threshs are floats

+++ src/coco_kernels.cu
@@ -32,7 +32,7 @@ static image det_s;
 static image disp ;
 static cv::VideoCapture cap;
 static float fps = 0;
-static int demo_thresh = 0;
+static float demo_thresh = 0;
 
 void *fetch_in_thread_coco(void *ptr)
 {
